# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Hifive football api",
    "version": "1.0.0",
    "title": "HifiveFootball",
    "termsOfService": "http://api.hifivefootball.com/",
    "contact": {
      "email": "ddastudio@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.hifivefootball.com:5300",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/contents": {
      "post": {
        "tags": [
          "contents"
        ],
        "summary": "content",
        "description": "post content",
        "operationId": "postContent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "post board  //form data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostContentBody"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContentHeader"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "contents"
        ],
        "summary": "content",
        "description": "update content",
        "operationId": "putContent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "description": "user authorization",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "post board  //form data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PutContentBody"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContentHeader"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/contents/{id}": {
      "get": {
        "tags": [
          "contents"
        ],
        "summary": "Finds Pets by status",
        "description": "Multiple status values can be provided with comma separated strings",
        "operationId": "getContent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "content id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ApiResponse"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "array",
              "items": {
                "type": "null"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ContentHeader": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "arena_id": {
          "type": "integer"
        },
        "deleted": {
          "type": "integer"
        },
        "reported": {
          "type": "integer"
        },
        "comments": {
          "type": "integer"
        },
        "scraps": {
          "type": "integer"
        },
        "board_id": {
          "type": "integer"
        },
        "preview": {
          "type": "string"
        },
        "likers": {
          "type": "integer"
        },
        "tags": {
          "type": "string"
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        },
        "bodytype": {
          "type": "integer"
        },
        "celltype": {
          "type": "integer"
        },        
        "link": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_action": {
          "$ref": "#/definitions/UserAction"
        }
      },
      "xml": {
        "name": "ContentHeader"
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "avatar_url": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserAction": {
      "type": "object",
      "properties": {
        "like": {
          "type": "integer"
        },
        "scrap": {
          "type": "integer"
        },
        "hasImgs": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "user_action"
      }
    },
    "PostContentBody": {
      "type": "object",
      "properties": {
        "board_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "preview": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "tags": {
          "type": "string"
        },
        "imgs": {
          "type": "string"
        },
        "bodytype": {
          "type": "integer",
          "enum": [
            1,
            2,
            3
          ]
        },
        "celltype": {
          "type": "integer"
        },         
        "allowcomment": {
          "type": "integer"
        },
        "link": {
          "type": "string"
        }
      },
      "xml": {
        "name": "PostContent"
      }
    },
    "PutContentBody": {
      "type": "object",
      "properties": {
        "content_id": {
          "type": "integer"
        },
        "board_id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "preview": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "tags": {
          "type": "string"
        },
        "imgs": {
          "type": "string"
        },
        "bodytype": {
          "type": "integer",
          "enum": [
            1,
            2,
            3
          ]
        },
        "celltype": {
          "type": "integer"
        },         
        "allowcomment": {
          "type": "integer"
        },
        "link": {
          "type": "string"
        }
      },
      "xml": {
        "name": "PutContent"
      }
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  },
...
}